centralise stat into the StatManager Class
