+++ Assets/Scripts/Models/Entities/Factions/Ally.cs
@@ -5,6 +5,6 @@ public abstract class Ally : Fighter
     protected override void Die()
     {
         base.Die();
-        KillCounter.IncrementAllyDeathCount();
+        StatManager.IncrementAllyDeathCount();
     }
 }
