+++ Assets/Scripts/stats/StatManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fb5a32222f7bf4742b66f62a85e34cac
+guid: f55090ed5a6ecbc49a7a9e0ea5552266
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
